rename lib to internal.
